factor out parse bitmap and color fn
