+++ src/functions/color.js
@@ -1,3 +1,5 @@
+import tinyColor from 'tinycolor2';
+
 import { isNodePattern, throwError } from '../utils/error-checking';
 import * as constants from '../constants';
 import { mulTable, shgTable } from './blur-tables';
@@ -876,3 +878,88 @@ export function displace(map, offset, cb) {
 
     return this;
 }
+
+/**
+ * Apply multiple color modification rules
+ * @param actions list of color modification rules, in following format: { apply: '<rule-name>', params: [ <rule-parameters> ]  }
+ * @param (optional) cb a callback for when complete
+ * @returns this for chaining of methods
+ */
+function colorFn(actions, cb) {
+    if (!actions || !Array.isArray(actions)) {
+        return throwError.call(this, 'actions must be an array', cb);
+    }
+
+    const originalScope = this;
+    this.scanQuiet(0, 0, this.bitmap.width, this.bitmap.height, function(
+        x,
+        y,
+        idx
+    ) {
+        let clr = tinyColor({
+            r: this.bitmap.data[idx],
+            g: this.bitmap.data[idx + 1],
+            b: this.bitmap.data[idx + 2]
+        });
+
+        const colorModifier = function(i, amount) {
+            const c = clr.toRgb();
+            c[i] = Math.max(0, Math.min(c[i] + amount, 255));
+            return tinyColor(c);
+        };
+
+        actions.forEach(action => {
+            if (action.apply === 'mix') {
+                clr = tinyColor.mix(clr, action.params[0], action.params[1]);
+            } else if (action.apply === 'tint') {
+                clr = tinyColor.mix(clr, 'white', action.params[0]);
+            } else if (action.apply === 'shade') {
+                clr = tinyColor.mix(clr, 'black', action.params[0]);
+            } else if (action.apply === 'xor') {
+                const clr2 = tinyColor(action.params[0]).toRgb();
+                clr = clr.toRgb();
+                clr = tinyColor({
+                    r: clr.r ^ clr2.r,
+                    g: clr.g ^ clr2.g,
+                    b: clr.b ^ clr2.b
+                });
+            } else if (action.apply === 'red') {
+                clr = colorModifier('r', action.params[0]);
+            } else if (action.apply === 'green') {
+                clr = colorModifier('g', action.params[0]);
+            } else if (action.apply === 'blue') {
+                clr = colorModifier('b', action.params[0]);
+            } else {
+                if (action.apply === 'hue') {
+                    action.apply = 'spin';
+                }
+
+                const fn = clr[action.apply];
+
+                if (!fn) {
+                    return throwError.call(
+                        originalScope,
+                        'action ' + action.apply + ' not supported',
+                        cb
+                    );
+                }
+
+                clr = fn.apply(clr, action.params);
+            }
+        });
+
+        clr = clr.toRgb();
+        this.bitmap.data[idx] = clr.r;
+        this.bitmap.data[idx + 1] = clr.g;
+        this.bitmap.data[idx + 2] = clr.b;
+    });
+
+    if (isNodePattern(cb)) {
+        return cb.call(this, null, this);
+    }
+
+    return this;
+}
+
+export const color = colorFn;
+export const colour = colorFn;
